backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
pandas==2.3.3
openai>=1.3.7,<2.0.0
python-dotenv==1.2.2
pandasai==3.0.0
pandasai-openai==0.1.6
pandasai-litellm==0.0.1
xlrd==2.0.2
openpyxl==3.1.5
python-multipart==0.0.22